restored edgetpumodel.py orig file. resiliency improved
